+++ src/test/java/fashionable/simba/yanawaserver/acceptance/AuthAcceptanceTest.java
@@ -166,35 +166,4 @@ class AuthAcceptanceTest extends AcceptanceTest {
         ExtractableResponse<Response> response = 로그인_발급_요청("12", "password-admin");
         assertThat(response.statusCode()).isEqualTo(HttpStatus.UNAUTHORIZED.value());
     }
-
-
-    /**
-     * When Refresh Token을 입력하면 Access Code을 발급 받고
-     * Then Access Code로 Access Token과 Refresh Token을 재발급 받고
-     * Then Access Token으로 사용자 정보를 조회한다.
-     */
-    @Test
-    void refresh_token() {
-        // given
-        String id = getId("admin");
-
-        String accessCode = 로그인_코드_발급(id, PASSWORD_ADMIN);
-        String refreshToken = 로그인_요청(accessCode).jsonPath().getString("refreshToken");
-
-        ExtractableResponse<Response> response = 코드_재갱신_요청(refreshToken);
-
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-        assertThat(response.jsonPath().getString("accessToken")).isNotNull();
-    }
-
-    /**
-     * When 유효하지 않은 Refresh Token을 입력하면
-     * Then 401 예외가 발생한다.
-     */
-    @Test
-    void refresh_token_failed() {
-        ExtractableResponse<Response> response = 코드_재갱신_요청("invalid refreshToken");
-
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.UNAUTHORIZED.value());
-    }
 }
