+++ src/test/java/fashionable/simba/yanawaserver/documentation/ExpiredDocumentation.java
@@ -2,6 +2,10 @@ package fashionable.simba.yanawaserver.documentation;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+
+import java.util.HashMap;
+import java.util.Map;
 
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.preprocessRequest;
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.preprocessResponse;
@@ -9,6 +13,23 @@ import static org.springframework.restdocs.operation.preprocess.Preprocessors.pr
 import static org.springframework.restdocs.restassured3.RestAssuredRestDocumentation.document;
 
 public class ExpiredDocumentation extends Documentation {
+    @Test
+    void refreshToken() {
+        Map<String, String> params = new HashMap<>();
+        params.put("refreshToken", refreshToken);
+
+        givenNotOauth()
+            .filter(document("token/refresh",
+                preprocessRequest(prettyPrint()),
+                preprocessResponse(prettyPrint())))
+            .contentType(MediaType.APPLICATION_JSON_VALUE)
+            .body(params)
+            .when().post("token/refresh")
+            .then().log().all()
+            .statusCode(HttpStatus.OK.value())
+            .extract();
+    }
+
     @Test
     void accessToken_expired() {
         givenOauth()
@@ -22,7 +43,7 @@ public class ExpiredDocumentation extends Documentation {
     }
 
     @Test
-    void logout() {
+    void refreshToken_expired() {
         givenOauth()
             .filter(document("expired/refresh",
                 preprocessRequest(prettyPrint()),
